DABOM-200 refactor: component 폴더 내 ui 폴더 제거
